% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.challenges_all}
\alias{get.challenges_all}
\title{Gets all the challenges from PoH Graph}
\usage{
get.challenges_all(api_url, method)
}
\arguments{
\item{api_url}{the api address}

\item{method}{the method to return the results}
}
\value{
Depends on the methods used.
}
\description{
Gets all the challenges from PoH Graph
}
\examples{
get.challenges_all("https://api.thegraph.com/subgraphs/name/kleros/proof-of-humanity-mainnet", "json1")
}
