% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.dailyUBIUSD}
\alias{get.dailyUBIUSD}
\title{Gets daily UBI/USD values from UniSwap v2}
\usage{
get.dailyUBIUSD(daynumbers)
}
\arguments{
\item{daynumber}{an integer or a vector of day numbers. Day 0 is March 11,
2021}
}
\value{
An object data.frame with date as POSIXct and priceUSD as numeric.
}
\description{
Gets daily UBI/USD values from UniSwap v2
}
\examples{
get.dailyUBIUSD(seq(0:10))
}
