% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get.txns_all}
\alias{get.txns_all}
\title{Gets all transactions from a contract address}
\usage{
get.txns_all(contract_address)
}
\arguments{
\item{contract_address}{a character value of the contract address}
}
\value{
An object data.frame
}
\description{
Gets all transactions from a contract address
}
\examples{
get.txns_all("0xdd1ad9a21ce722c151a836373babe42c868ce9a4")
}
