% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{getRegistry}
\alias{getRegistry}
\title{Gets gets all the submissions in the registry}
\usage{
getRegistry()
}
\value{
A dataframe with all submissions, including removed ones.
}
\description{
Gets gets all the submissions in the registry
}
\examples{
getRegistry()
}
